Move everything to main index.tsx
