+++ src/types.tsx
@@ -7,6 +7,7 @@ import {
   TextInputFocusEventData,
   ColorValue,
 } from 'react-native';
+import React from 'react';
 
 export type SearchBarCommands = {
   focus: () => void;
